[.NET 5] fix warning for $(TFV) of 5.0
In our current .NET 5 builds, you always hit:

    warning XA1008: The TargetFrameworkVersion (Android API level 21) is lower than the targetSdkVersion (29).
    Please increase the `$(TargetFrameworkVersion)` or decrease the `android:targetSdkVersion` in the `AndroidManifest.xml` so that the API levels match.

It is giving a warning for Google Play, because
`$(TargetFrameworkVersion)` is equal to 5.0.

To fix this, I modified the `<CheckGoogleSdkRequirements/>` MSBuild
task so that `TargetFrameworkVersion` is optional. For .NET 5 builds,
we can pass in `$(_AndroidApiLevel)` instead.

This way we will still get some important warnings for .NET 5 such as
`targetSdkVersion` or `minSdkVersion` that need to be changed in the
`AndroidManifest.xml` file.

I also fixed up `LastBuildOutput`, we needed the value for .NET 5
MSBuild tests and I removed unnecessary usage of `yield return`.