[.NET 5] default to --self-contained=true (#4587)
Currently with a .NET 5 build you have to do:

	dotnet publish foo.csproj --self-contained

Or:

	dotnet publish foo.csproj -p:SelfContained=true

On Android platforms, it doesn't make sense to have a non-self-
contained app.  That would mean there is a system-wide .NET 5
installation, which is not going to happen on Android.

In .NET 5 running on Windows/MacOS, this means the application relies
on the BCL living in a system-wide directory somewhere.

To simplify the command-line arguments we should default
`$(SelfContained)` to `true` when it is an Android application.