Save apkdesc files to output path (#4591)
Changes: https://github.com/radekdoulik/apkdiff/compare/dc6cfc...cb0083b9

  * radekdoulik/apkdiff@cb0083b: Fix indentation, put the XA link  on separate line
  * radekdoulik/apkdiff@092e40a: Added links to the upstream repo and XA location
  * radekdoulik/apkdiff@cdb7593: Added DEX comparison example
  * radekdoulik/apkdiff@17f242b: Update usage info in README.md
  * radekdoulik/apkdiff@1dbdb00: Added options to specify path of saved description
  * radekdoulik/apkdiff@e6390f6: Initial DEX comparison
  * radekdoulik/apkdiff@7c8d5d8: Dispose PEReaders

Mostly to have the options to save descriptions to specified path.
Also contains the `.dex` file comparison.

Add the configuration and the flavor to the filename, and save them to
`$(OutputPath)`.  This way we can just copy the files when we need to
update reference `.apkdesc` files.