[AndroidClientHandler] Wrap Java exceptions in .NET ones
Context: 60363ef45cd6994a322adac1b73d4271b0658c1e

This commit a follow-up to 60363ef4 in that it causes less Java
exceptions to "leak" to applications which use code shared between
platforms that might not be able to check for Java exception types
directly (because they don't link against `Mono.Android.dll`).  However,
instead of making the behavior optional, as implemented in 60363ef4,
this commit unconditionally wraps `Java.IO.Exception` and some of its
derivatives in the standard .NET `WebException`.  This makes the handler
to behave in a manner that's closer to how other `HttpClient` custom
handlers behave.

The `$(AndroidBoundExceptionType)` MSBuild property must be set to
`System` in order to use the wrapped exceptions.