[tests] ignore .NET 5 tests (#4603)
Some of our tests are failing on CI with:

	dotnet\sdk\5.0.100-preview.2.20176.6\Sdks\Microsoft.NET.Sdk\targets\Microsoft.NET.Sdk.FrameworkReferenceResolution.targets(329,5):
	error NETSDK1082: There was no runtime pack for Microsoft.NETCore.App available for the specified RuntimeIdentifier 'android.21-x86'.

These tests work fine if you install a master build of .NET 5, but
not with the current .NET 5 Preview 3 release.  We will need a build
of .NET 5 Preview 4 for these tests to work.

Let's ignore them for now, so our CI will be more greenish.