[.NET 5] add a `Run` MSBuild target
Currently with .NET 5, you can build, deploy, and run an app with:

    dotnet publish foo.csproj -t:Install -t:StartAndroidActivity

This doesn't match the behavior of other platforms that use `dotnet
run` to launch.

Unfortunately `dotnet run` won't work in its current state:

    Unable to run your project.
    Ensure you have a runnable project type and ensure 'dotnet run' supports this project.
    A runnable project should target a runnable TFM (for instance, netcoreapp2.0) and have OutputType 'Exe'.
    The current OutputType is 'Library'.

But we can supply a `Run` target to start Android applications:

https://github.com/dotnet/sdk/blob/645919b8e850771a73f2f7a31b97038d6bdf1092/src/Tasks/Microsoft.NET.Build.Tasks/targets/Microsoft.NET.Sdk.targets#L755-L758

This at least allows this command to work:

    dotnet publish foo.csproj -t:Run

I updated a test to use this command as well.