[xaprepare] Don't require binaries when unable to install them (#4597)
`xaprepare` defines a set of binaries/programs which are deemed
essential and, thus, it enforces their presence before the scenario
runs.  However, there can be scenarios (e.g. `-s:UpdateMono`) which
disallow installation of any programs and should any of the essential
programs be missing, we'd see a failure similar to:

	Required program '7za' could not be found
	System.InvalidOperationException: Required program '7za' could not be found
	  at Xamarin.Android.Prepare.OS.Which (System.String programPath, System.Boolean required)
	  at Xamarin.Android.Prepare.EssentialTools.Init (Xamarin.Android.Prepare.Context context)
	  at Xamarin.Android.Prepare.Context+<Init>d__201.MoveNext ()
	--- End of stack trace from previous location where exception was thrown ---
	  at System.Runtime.CompilerServices.TaskAwaiter.ThrowForNonSuccess (System.Threading.Tasks.Task task)
	  at System.Runtime.CompilerServices.TaskAwaiter.HandleNonSuccessAndDebuggerNotification (System.Threading.Tasks.Task task)
	  at System.Runtime.CompilerServices.TaskAwaiter.ValidateEnd (System.Threading.Tasks.Task task)
	  at System.Runtime.CompilerServices.TaskAwaiter`1[TResult].GetResult ()
	  at Xamarin.Android.Prepare.App+<Run>d__3.MoveNext ()

Allow such scenarios to run in the broken environment by making it
possible to ignore certain missing programs.  Scenarios which
disallow installation of any programs must be very careful in what
programs they use.