[Xamarin.Android.Build.Tasks] Fixup the AndroidManifest.xml path.
Fixes: http://work.devdiv.io/662837

During the build process, the `$(AndroidManifest)` file -- usually
`Properties\AndroidManifest.xml` -- is "merged" with generated XML
fragments based on custom attributes such as
`Android.App.ActivityAttribute`.

If there is an error in the generated file -- often because of an
error within `$(AndroidManifest)`, e.g. an invalid reference to an
Android resource -- then `aapt`/`aapt2`/etc. will emit an error message
which mentions `$(IntermediateOutputPath)android\AndroidManifest.xml`:

	…\obj\Debug\MonoAndroid70\android\manifest\AndroidManifest.xml(7): error APT2258: resource mipmap/invalid (aka com.xamarin.android.helloworld:mipmap/invalid) not found

Unfortunately, there's no easy way to accurately determine if the cause
of the error message is due to `$(AndroidManifest)` or instead due to
the *generated* portions.  Is the error in the customer's code, or in
*Xamarin.Android* code?

Update the error message to use `$(AndroidManifest)` as the "source" of
the error, not something in `$(IntermediateOutputPath)`, and to also
mention that there could be multiple causes of the error:

	…\Properties\AndroidManifest.xml(7): error APT2258:
	resource mipmap/invalid (aka com.xamarin.android.helloworld:mipmap/invalid) not found.
	This error is likely caused by an issue with the AndroidManifest.xml file or an Android
	manifest generation attribute in a source code file.

Additionally, add an XML comment to the top of generated
`AndroidManifest.xml` file, indicating that the file is generated:

	<!--
	  This code was generated by a tool.
	  It was generated from {0}
	  Changes to this file may cause incorrect behavior and will be lost if
	  the contents are regenerated.
	  -->