Fix build break on newly reimaged bot (#4607)
In the process of creating a new macOS OSS PR workflow, I encountered
an issue where the build would consistently break when executing
`make prepare-update-mono`.  The break occurred on the following
command *on a newly re-imaged agent (bot)*:

    pgrep -lfi VBCSCompiler.exe

The command returns 0 if the managed compiler server `VBCSCompiler.exe`
is running; otherwise, it returns 1 if the process is not running.
When executing the build pipeline on the newly re-imaged machine, the
process is not running, `pgrep` command returns 1, and the build fails
with the following error: 

	pgrep -lfi VBCSCompiler.exe
	make: *** [prepare-update-mono] Error 1

Turns out that on established build machines in the pool,
`VBCSCompiler.exe` is consistently found to be running at this stage
of the build.  If you inspect a build log, you will see that the
`pgrep` command logged that `VBCSCompiler.exe` is running.  Here's an
example from a recent build log:

	pgrep -lfi VBCSCompiler.exe
	36187 /Library/Frameworks/Mono.framework/Versions/6.10.0/bin/mono-sgen64 /Library/Frameworks/Mono.framework/Versions/6.10.0/lib/mono/msbuild/Current/bin/Roslyn/VBCSCompiler.exe -pipename:sX2CFXy8VTADY0R+V+y_1o2brmCa+1ch6aJTNiSJU4w

In discussing the issue with @directhex, the intent of the command is
quite the opposite: the intent is to ensure that `VBCSCompiler.exe` is
not running at this point of time.  Based on the purpose of the target
to perform a mono update, we should ensure that processes associated
with the `mono` toolset are *shut down* and not running prior to the
update.  Since `VBCSCompiler.exe` is part of the `mono` toolset it
seems to make more logical sense to ensure that it is not running.  In
other words, ensure that the next invocation of `MSBuild` starts the
compiler server from the new location if `mono` is updated.