[monodroid, Xamarin.Android.Build.Tasks] Interpreter support (#4618)
Context: https://www.mono-project.com/news/2017/11/13/mono-interpreter/

Add *experimental* support to use [Mono's IL interpreter][0] instead
of using the JIT or (Hybrid)AOT backends by setting the
`$(_AndroidUseInterpreter)` MSBuild property to True.

This is a proof-of-concept to explore how the interpreter backend
performs compared to the JIT backend at various tasks.  It is not
currently intended for production use.

Note that Mono's IL interpreter does *not* currently support running
on x86 targets.  Issue an XA0124 error if the interpreter is enabled
when targeting x86 devices.

One "interesting" -- but not explored -- result is that startup time
is [*reduced* when using the interpreter][1]; on a Pixel 3 XL launching
a arm64-v8a build of `tests/Xamarin.Forms-Performance-Integration`:

| Description | native-to-managed | Runtime.init | ActivityDisplayed |
|------------:|------------------:|-------------:|------------------:|
| JIT         |         53.660 ms |    92.821 ms |        1091.30 ms |
| Interpreter |         32.019 ms |   112.288 ms |         897.50 ms |

Here, native-to-managed time and ActivityDisplayed times are reduced,
while `Runtime.init()` time is increased.

We did not explore why this would be the case, or what other
performance changes would be present after app startup.

[0]: https://www.mono-project.com/news/2017/11/13/mono-interpreter/
[1]: https://gist.github.com/grendello/3510f2c0621eb360cd063d6c00b4b1e8#file-results-md