[Microsoft.Android.Sdk] change dotnet publish/build behavior (#4636)
Currently, we are producing `.apk` files with .NET 5 with:

	dotnet publish app.csproj

The problem with this is that `dotnet build` needs to produce
something that is "runnable".  `dotnet build` is the developer loop,
while `dotnet publish` is something you run for a release.

Ideally, to produce an `.apk` you would instead do:

	dotnet build app.csproj

`dotnet publish` would be used for distribution to Google Play.
Our first implementation would likely support re-signing the app and
copying the output to a local `publish` directory.

For now, I've removed our usage of `$(_CorePublishTargets)` and moved
these targets to `$(BuildDependsOn)`.  The benefit here, is we are no
longer using any private MSBuild properties for .NET 5.

I updated our tests in these areas. I dropped the tests running
`dotnet publish`, as they are no longer valid.

Down the road:

  * Make `dotnet publish` do *something*.

    Once we know what it should *do*.

  * Implement `dotnet run`.