[.NET 5] support for binding projects
The current idea for binding projects in .NET 5 is they won't be a
distinct project type at all anymore. It would be completely valid to
put this in a .NET 5 Xamarin.Android app with `$(OutputType)` of
`Exe`:

    <InputJar Include="foo.jar" />

A project with an `$(OutputType)` of `Library` would likely do:

    <EmbeddedJar Include="foo.jar" />

To make this work, I've split up `Xamarin.Android.Bindings.targets`
into a few different files:

* `Xamarin.Android.Bindings.Core.targets` - the targets shared between
  .NET 5 and "legacy" binding projects. It imports the other new files.
* `Xamarin.Android.Bindings.ClassParse.targets` - is only imported
  when `$(AndroidClassParser)` is `class-parse`.
* `Xamarin.Android.Bindings.JarToXml.targets` - is only imported
  when `$(AndroidClassParser)` is `jar2xml`.
* `Xamarin.Android.Bindings.Documentation.targets` - is imported by
  "legacy" binding projects, contains targets related to documentation
  extraction/generation. We will eventually use this in .NET 5 when
  `mdoc.exe` is available or we have an alternative.

Once these were split up, to get bindings working in .NET 5 projects I
had to:

1. Import `Xamarin.Android.Bindings.Core.targets`.
2. Update `Microsoft.Android.Sdk.BuildOrder.targets` to include the
   binding-specific targets.
3. Make the binding targets only run if there are any `@(InputJar)` or
   `@(EmbeddedJar)`.

One hangup is that bindings require:

    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>

We probably don't want this set for all projects by default. I set
this if any of the binding targets run, so that once you add a
`@(InputJar)` it gets turned on.

I updated `XASdkTests` to verify binding `.jar` files works in a
library project and an app project.

Other changes:

* Added a new `AndroidDotnetToolTask` for swapping between
  `generator.exe` and `dotnet generator.dll` calls.
* I ported `javadoc2mdoc.csproj` to use short-form projects, and
  multi-target both `net472` and `netcoreapp3.1`.
* The `<ImportJavaDoc/>` now extends `AndroidDotnetToolTask`.
* `$(ApplicationResolveReferencesDependsOn)` is unused, so I removed
  usage of it.
* I removed the `_ResolveMonoAndroidFramework` target, it was not
  needed.

TODO:

* Item groups like `@(JavaSourceJar)` can't fully work yet in .NET 5.
  `Xamarin.Android.Documentation.targets` uses `mdoc.exe`, which won't
  run under `dotnet` yet.
* For now, `Xamarin.Android.Documentation.targets` is not imported in
  .NET 5 projects.