Fix binary_search
Commit af13997b asked, "is binary_search broken?"

Answer: Yes, because 5016dc40 broke it:

@@ -149,8 +149,8 @@ EmbeddedAssemblies::binary_search (const Key *key, const Entry *base, size_t nme
 	}

 	constexpr size_t size = sizeof(Entry);
+	const Entry *ret = nullptr;
 	while (nmemb > 0) {
-		const Entry *ret;
 		if constexpr (use_extra_size) {
 			ret = reinterpret_cast<const Entry*>(reinterpret_cast<const uint8_t*>(base) + ((size + extra_size) * (nmemb / 2)));
 		} else {
@@ -168,11 +168,29 @@ EmbeddedAssemblies::binary_search (const Key *key, const Entry *base, size_t nme
 			}
 			nmemb -= nmemb / 2 + 1;
 		} else {
-			return ret;
+			break;
 		}
 	}

The bug is that in moving `ret` up a scope, nothing *clears out* `ret`
on error!  Thus, it's "garbage" when the `while(nmemb>0)` loop
terminates and no key has been found, which is why it *always*
returned a non-null value.

(Rocket science!  And bad code refactoring.  And PEBKAC.)

The fixis to `ret=nullptr` on every loop iteration unless we have a
value value.