[Xamarin.Android.Build.Tasks] simplify d8/r8 multi-dex support
When `android:minSdkVersion` is 21 or higher, d8 can automatically
support multi-dex. It will merely generate multiple `classes.dex`
files as needed, and we won't have a need for an
`$(AndroidEnableMultiDex)` MSBuild property.

Changes when `$(AndroidDexTool)` is set to `d8`:

* We no longer use `android-support-multidex.jar`.
* We no longer use `android.support.multidex.MultiDexApplication`.
* `$(AndroidEnableMultiDex)` is not used for anything.

These changes are not made when `AndroidDexTool=dx`

This makes the `<R8/>` MSBuild task much simpler, as many of the
incoming values are no longer needed. `<R8/>` will now only be needed
when `AndroidLinkTool=r8` and developers are opting into code
shrinking.