Fix <CreateMultiDexMainDexClassList/>?
Context: https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=3729696&view=ms.vss-test-web.build-test-results-tab&runId=13228096&resultId=100066&paneView=attachments

Current failures?  `<CreateMultiDexMainDexClassList/>` is unhappy.
More precisely, JDK11 `java` is unhappy:

	$ /Users/runner/Library/Android/jdk/bin/java \
		-Djava.ext.dirs=/Users/runner/Library/Android/sdk/build-tools/29.0.2/lib \
		com.android.multidex.MainDexListBuilder …
	-Djava.ext.dirs=/Users/runner/Library/Android/sdk/build-tools/29.0.2/lib is not supported.  Use -classpath instead.
	CREATEMULTIDEXMAINDEXCLASSLIST : error : Could not create the Java Virtual Machine.
	CREATEMULTIDEXMAINDEXCLASSLIST : error : A fatal exception has occurred. Program will exit.
	The command exited with code 1.

JDK11 `java` no longer supports `-Djava.ext.dirs`.

Update `<CreateMultiDexMainDexClassList/>` so that it constructs a
`java -classpath` value and uses that.

Update `$(AndroidSdkBuildToolsVersion)` to 30.0.0-rc4

TODO: Is the `java -classpath` quoting correct?

TODO: So about the `$(AndroidSdkBuildToolsVersion)` semantics: should
it be in any way correlated with the JDK version in use?

If so, *how*?

We *know* that if `$(AndroidSdkBuildToolsVersion)` is >= 30, we need
JDK11.  How can we check that?

If `$(AndroidSdkBuildToolsVersion)` is < 30, does the JDK matter?
Given that all current failures are because of `java -Djava.ext.dirs`,
it migiht not matter, i.e JDK 11 works with 29.0.2.