[monodroid] Set cmake path to include build type
The `CMAKE_LIBRARY_OUTPUT_DIRECTORY` value was same for Debug and
Release build types (which is not the same as Configuration, we build
both types in each configuration). And because we configure these
in parallel, it might result in concurent file writes.

I think this is causing these errors on build bots:

    error : CMake Error at C:/Users/dlab14/android-toolchain/sdk/cmake/3.10.2.4988404/share/cmake-3.10/Modules/CheckCXXSourceCompiles.cmake:91 (file): [C:\A\vs2019xam00000E-1\_work\2\s\src\monodroid\monodroid.csproj]
    error :   file failed to open for writing (Permission denied): [C:\A\vs2019xam00000E-1\_work\2\s\src\monodroid\monodroid.csproj]
    error :     C:/A/vs2019xam00000E-1/_work/2/s/src/monodroid/obj/Release/x86_64-Release/CMakeFiles/CMakeTmp/src.cxx [C:\A\vs2019xam00000E-1\_work\2\s\src\monodroid\monodroid.csproj]