[.NET 5] Update TargetFramework to net5.0 (#4692)
Bumps to .NET 5.0.100-preview.6.20265.2

The latest .NET 5 previews now have support for a $(TargetFramework) of
`net5.0`, which allows us to replace our usage of `netcoreapp5.0` with
`net5.0`.

The .NET 5 version of `_ResolveSdks` has also been updated to avoid
avoid usage of $(TargetFramework) when setting the path to
`Mono.Android.dll`.  This fixes an issue that was occurring when using
the latest .NET 5 nightly preview, which was the result of Android
projects using targeting packs with different $(TargetFramework) values.

    error XARSD7004: System.ArgumentException: `/usr/local/share/dotnet/packs/Microsoft.NETCore.App.Ref/5.0.0-preview.6.20264.1/ref/netcoreapp5.0` must be a directory! (Parameter 'frameworkDirectories')
    error XARSD7004:    at Xamarin.Android.Tools.AndroidVersions..ctor(IEnumerable`1 frameworkDirectories) in /Users/peter/source/pj/xamarin-android/external/xamarin-android-tools/src/Xamarin.Android.Tools.AndroidSdk/AndroidVersions.cs:line 30
    error XARSD7004:    at Xamarin.Android.Tasks.MonoAndroidHelper.RefreshSupportedVersions(String[] referenceAssemblyPaths)
    error XARSD7004:    at Xamarin.Android.Tasks.ResolveSdks.RunTask()
    error XARSD7004:    at Xamarin.Android.Tasks.AndroidTask.Execute()

* Use targeting pack metadata to set $(_XATargetFrameworkDirectories)

Certain tasks require that $(_XATargetFrameworkDirectories) contains
paths to both Mono.Android.dll and mscorlib.dll. In order to ensure that
this property contains all known assembly paths we'll use the
@(ResolvedTargetingPack) ItemGroup. These items contain metadata
definitions for %(TargetFramework) and %(PackagePath), which can be
combined to construct the full assembly path for each pack.