[Xamarin.Android.Build.Tasks] `tools:node="remove"` in Manifest (#4682)
Fixes: https://github.com/xamarin/xamarin-android/issues/4662

`manifestmerger.jar` did not work in the way we expeected.  Its primary
purpose is to allow users to remove/modify manifest items from
*LIBRARY* projects.  What it doesn't do it process items in the main
manifest itself.

As part of our build system we sometimes auto generate manifest items
based on C# attributes within a referenced assembly.  A good example of
this is the `ZXing.Net.Mobile` packages, which implicitly adds a
`<uses-permission/>` element for `android.permission.CAMERA`.

In #4662 the user expectation was that if they add `tools:node="remove"`
to the `uses-permission` element in their own manifest, this would stop
the permission from being added.  Unfortunately this did NOT happen.
The `manifestmerger.jar` completely ignored this entry and it ended up
in the final manifest.  This is because the permission itself is NOT
coming from a library manifest file, but is being injected directly
into the main manifest file by our own code.

Work around this limitation and to get the expected user behaviour
by changing the `ManifestDocument` class to strip out all the elements
which contain a `tools:node="remove"` attribute.  This new
functionality will ONLY be called by the `ManifestMerger` class,
because that is the only configuration where `tools:node="remeove"`
will be used.